Sync test files moved components/sync/test/
Chromium change:
https://chromium.googlesource.com/chromium/src/+/c331eeb96a5b5b4c4a5c13121fe5f65e0b9290fa

commit c331eeb96a5b5b4c4a5c13121fe5f65e0b9290fa
Author: Victor Hugo Vianna Silva <victorvianna@google.com>
Date:   Mon Aug 15 20:13:37 2022 +0000

    [sync] Move components/sync/test/**/foo.* to components/sync/test/foo.*

    No behavior is changed, most changes are an automated application of
    tools/git/move_source_file.py. The only manual changes are:
    - DEPS files.
    - Changing #include to #import on iOS files to satisfy a presubmit.

    Bug: 948661
